requirements.txt
requests==2.31.0
google-cloud-storage==2.10.0
cloud-sql-python-connector==1.4.1
pymysql==1.1.0
pytz==2023.3
python-dotenv==1.0.0